+++ apps/desktop/src/i18n/en.ts
@@ -2954,190 +2954,4 @@ export const en: Translations = {
       toggle: open => `${open ? 'Show' : 'Hide'} sidebar`
     }
   },
-
-  kanban: {
-    nav: 'Kanban',
-    openBoard: 'Kanban: Open board',
-    countTip: (running, ready) => `Kanban — ${running} running, ${ready} ready`,
-    col: {
-      triage: { label: 'Triage', help: 'Raw ideas — a specifier fleshes out the spec.' },
-      todo: { label: 'Todo', help: 'Waiting on dependencies, or unassigned.' },
-      scheduled: { label: 'Scheduled', help: 'Waiting for a scheduled time to arrive.' },
-      ready: { label: 'Ready', help: 'Dependencies satisfied — assign a profile and the dispatcher runs it.' },
-      running: { label: 'Running', help: 'Claimed by a worker — an agent is on it. Set by the dispatcher.' },
-      blocked: { label: 'Blocked', help: 'The worker asked for human input.' },
-      review: { label: 'Review', help: 'A review agent is checking the work. Set by the dispatcher.' },
-      done: { label: 'Done', help: 'Completed; dependent children become ready.' },
-      archived: { label: 'Archived', help: 'Hidden from the default board view.' }
-    },
-    locked: {
-      review: 'Review is entered by the dispatcher when a review agent takes the card.',
-      running: 'Running is set by the dispatcher when a worker claims the card.',
-      scheduled: 'Scheduled needs a wake-up time — agents set it; it can’t be dragged into.'
-    },
-    arcRunning: 'An agent is working on this now.',
-    arcStale: 'Claimed, but no worker heartbeat for 2+ minutes — the dispatcher will reclaim it.',
-    title: 'Kanban',
-    orchestrationSettings: 'Orchestration settings',
-    newTask: 'New task',
-    filterCards: 'Filter cards…',
-    noMatch: 'No tasks match the filters',
-    noTasks: 'No tasks on this board',
-    open: 'Open',
-    select: 'Select (⌘-click)',
-    deselect: 'Deselect',
-    moveTo: label => `Move to ${label}`,
-    delete: 'Delete',
-    reviewChecking: 'A review agent is checking the completed work.',
-    attachedTip: name => `${name} is attached — the dispatcher hands this over on its next tick (≤1m).`,
-    orchestratorTip: name => `${name} (the orchestrator) picks this up on the next tick and writes the spec.`,
-    autoAssignTip: name => `Auto-assigns to “${name}” (kanban.default_assignee) on the next dispatch tick.`,
-    wontRun: "won't run",
-    wontRunTip:
-      'Ready cards only run once a profile is assigned. Open the card and set an assignee, or configure a default assignee in orchestration settings.',
-    noHeartbeat: 'no heartbeat',
-    expand: label => `Expand ${label}`,
-    collapse: label => `Collapse ${label}`,
-    newTaskIn: label => `New task in ${label}`,
-    empty: 'Empty',
-    unassigned: 'unassigned',
-    filters: 'Filters',
-    allProfiles: 'All profiles',
-    allTenants: 'All tenants',
-    showArchived: 'Show archived',
-    groupRunning: 'Group Running by profile',
-    nSelected: n => `${n} selected`,
-    moveToShort: 'Move to',
-    assign: 'Assign',
-    unassignAction: 'Unassign',
-    archive: 'Archive',
-    clearSelection: 'Clear selection (Esc)',
-    refused: 'refused',
-    bulkFailed: (failed, total, err) => `${failed} of ${total} failed — ${err}. Failed cards stay selected.`,
-    titlePlaceholderTriage: 'Rough idea — a specifier will flesh it out',
-    titlePlaceholder: 'Title',
-    descPlaceholder: 'Description (optional)',
-    priority: 'Priority',
-    workspace: 'Workspace',
-    boardDefaultSuffix: ' · board default',
-    workspaceOverride: 'Workspace path (optional override)',
-    workspaceInherit: 'Inherits the board’s project directory',
-    workspaceInheritDir: dir => `Leave empty to inherit ${dir}`,
-    workspaceInheritGeneric: 'Leave empty to inherit the board’s project directory.',
-    assignee: 'Assignee',
-    defaultOption: name => `${name} (default)`,
-    parkedOption: "unassigned (parked — won't run)",
-    skills: 'Skills (comma-separated)',
-    skillsPlaceholder: 'translation, github',
-    parent: "Parent (blocks until it's done)",
-    noParent: '— no parent —',
-    goalMode: "Goal mode (worker loops until a judge agrees it's done)",
-    creating: 'Creating…',
-    createTask: 'Create task',
-    cancel: 'Cancel',
-    save: 'Save',
-    estimate: 'Estimate',
-    estimateEffort: 'Estimate effort',
-    estimating: 'Estimating…',
-    reEstimate: 'Re-estimate',
-    makesModelCall: 'makes a model call',
-    estimateTip: 'Rough token + complexity estimate from the auxiliary model — makes a model call.',
-    estimateTipLong: 'Runs a quick auxiliary-model call to estimate tokens + complexity. A rough guide, not a bill.',
-    roughEstimate: 'Rough estimate',
-    tokUnit: 'tok',
-    couldNotEstimate: 'Could not estimate',
-    complexity: { S: 'Small', M: 'Medium', L: 'Large' },
-    introBody:
-      'You don’t run the cards — agents do. Put a card in Ready with an assignee and an agent picks it up within a minute. No assignee, no run. Triage: an agent rewrites the idea into a proper task first. Todo: waiting on other cards. Scheduled: waiting on a timer. Running and Review: the agents’ lanes, hands off. Blocked: it’s waiting on you. Results come back on the card.',
-    introGotIt: 'Got it',
-    evtCreated: (where, assignee) => `created${where ? ` in ${where}` : ''}${assignee ? ` · assigned to ${assignee}` : ''}`,
-    evtMovedTo: col => `moved to ${col}`,
-    evtParentReopened: parent => `parent ${parent} reopened`,
-    evtAssignedTo: assignee => `assigned to ${assignee}`,
-    evtUnassigned: 'unassigned',
-    evtCommentBy: author => `comment by ${author}`,
-    evtClaimedReview: 'claimed by a review agent',
-    evtClaimedWorker: 'claimed by a worker',
-    evtWorkerStarted: 'worker started',
-    evtCompleted: 'completed',
-    evtBlocked: 'blocked — needs human input',
-    evtUnblocked: col => `unblocked${col ? ` → ${col}` : ' → Ready'}`,
-    evtReclaimed: 'reclaimed — returned to the queue',
-    evtSpecified: 'spec written by the triage agent',
-    evtPromoted: 'dependencies done — promoted to Ready',
-    evtScheduled: 'scheduled for later',
-    evtArchived: 'archived',
-    evtReprioritized: priority => `priority set to ${priority}`,
-    someone: 'someone',
-    metaPriority: 'Priority',
-    metaTenant: 'Tenant',
-    metaCreatedBy: 'Created by',
-    metaCreated: 'Created',
-    metaWorkerPid: 'Worker pid',
-    readyUnassignedTitle: 'Ready, but unassigned — this card will never run.',
-    readyUnassignedBody:
-      'The dispatcher only claims Ready cards that have an assignee. Pick a profile in the Assignee field above (or set a default assignee in the orchestration settings) and it runs within a minute.',
-    diagnosticsN: n => `Diagnostics · ${n}`,
-    commandCopied: 'Command copied',
-    description: 'Description',
-    editDescription: 'Edit description',
-    cancelEdit: 'Cancel edit',
-    noDescription: 'No description yet.',
-    result: 'Result',
-    latestSummary: 'Latest summary',
-    dependencies: 'Dependencies',
-    blockedBy: 'Blocked by',
-    blocks: 'Blocks',
-    comments: n => `Comments · ${n}`,
-    commentsHelpRunning:
-      'This task is running. Your note is folded into the worker’s current turn within a few seconds — no block/unblock dance. “Requeue with note” instead restarts the task from scratch with your note in context.',
-    commentsHelp:
-      'Comments are added to the task thread. When a worker picks the task up it reads them as part of its context.',
-    send: 'Send',
-    comment: 'Comment',
-    messageWorker: 'Message the running worker…',
-    addComment: 'Add a comment…',
-    deliveredLive: 'Delivered to the running worker within a few seconds.',
-    requeueWithNote: 'Requeue with note',
-    notePosted: 'Note posted — worker requeued',
-    activity: n => `Activity · ${n}`,
-    runs: n => `Runs · ${n}`,
-    workerLog: 'Worker log',
-    workerLogTail: 'Worker log · tail',
-    attachments: n => `Attachments · ${n}`,
-    noAttachments: 'No attachments yet.',
-    uploadAttachment: 'Upload attachment',
-    taskActions: 'Task actions',
-    copyTaskId: 'Copy task id',
-    copyTitle: 'Copy title',
-    copiedId: id => `Copied ${id}`,
-    copiedTitle: 'Copied title',
-    archiveTask: 'Archive task',
-    deleteTask: 'Delete task',
-    close: 'Close',
-    working: 'working',
-    board: 'Board',
-    newBoard: 'New board',
-    newBoardDots: 'New board…',
-    boardSettings: 'Board settings…',
-    boardSettingsFor: name => `Board settings — ${name}`,
-    name: 'Name',
-    boardNamePlaceholder: 'Board name',
-    slug: slug => `slug: ${slug}`,
-    project: 'Project',
-    noProject: 'No project (scratch sandboxes)',
-    projectHintPre:
-      'New tasks run in the project’s repo (a worktree per task); each task can still override its workspace at creation. Manage projects with ',
-    projectHintCmd: 'hermes project',
-    createBoard: 'Create board',
-    orchestratorProfile: 'Orchestrator profile',
-    defaultAssignee: 'Default assignee',
-    defaultParen: '(default)',
-    autoDecompose: 'Auto-decompose triage tasks',
-    profileDescriptions: 'Profile descriptions',
-    profileDescriptionsHint:
-      'Descriptions guide the decomposer’s routing. Auto-generate with the auxiliary model, or write your own.',
-    profileGoodAt: 'What is this profile good at?',
-    auto: 'Auto'
-  }
 }
