+++ apps/desktop/src/i18n/ja.ts
@@ -2799,190 +2799,4 @@ export const ja = defineLocale({
       toggle: open => `サイドバーを${open ? '表示' : '非表示'}`
     }
   },
-
-  kanban: {
-    nav: 'カンバン',
-    openBoard: 'カンバン: ボードを開く',
-    countTip: (running, ready) => `カンバン — 実行中 ${running}、待機 ${ready}`,
-    col: {
-      triage: { label: 'トリアージ', help: '生のアイデア — スペシファイアが仕様に整えます。' },
-      todo: { label: 'Todo', help: '依存関係の待ち、または未割り当て。' },
-      scheduled: { label: 'スケジュール', help: '予定時刻を待っています。' },
-      ready: { label: 'Ready', help: '依存関係が解決済み — プロフィールを割り当てるとディスパッチャが実行します。' },
-      running: { label: '実行中', help: 'ワーカーが取得済み — エージェントが作業中。ディスパッチャが設定します。' },
-      blocked: { label: 'ブロック', help: 'ワーカーが人間の入力を求めています。' },
-      review: { label: 'レビュー', help: 'レビューエージェントが作業を確認中。ディスパッチャが設定します。' },
-      done: { label: '完了', help: '完了。依存する子タスクが Ready になります。' },
-      archived: { label: 'アーカイブ', help: 'デフォルトのボード表示から非表示。' }
-    },
-    locked: {
-      review: 'レビューは、レビューエージェントがカードを取得するとディスパッチャによって設定されます。',
-      running: '実行中は、ワーカーがカードを取得するとディスパッチャによって設定されます。',
-      scheduled: 'スケジュールには起動時刻が必要です — エージェントが設定します。ドラッグでは移動できません。'
-    },
-    arcRunning: 'エージェントが現在作業中です。',
-    arcStale: '取得済みですが、2分以上ワーカーのハートビートがありません — ディスパッチャが再取得します。',
-    title: 'カンバン',
-    orchestrationSettings: 'オーケストレーション設定',
-    newTask: '新しいタスク',
-    filterCards: 'カードを絞り込み…',
-    noMatch: 'フィルタに一致するタスクはありません',
-    noTasks: 'このボードにタスクはありません',
-    open: '開く',
-    select: '選択（⌘クリック）',
-    deselect: '選択解除',
-    moveTo: label => `${label} へ移動`,
-    delete: '削除',
-    reviewChecking: 'レビューエージェントが完了した作業を確認中です。',
-    attachedTip: name => `${name} が担当 — ディスパッチャが次のティック（≤1分）で引き渡します。`,
-    orchestratorTip: name => `${name}（オーケストレーター）が次のティックでこれを取得し、仕様を書きます。`,
-    autoAssignTip: name => `次のディスパッチティックで「${name}」（kanban.default_assignee）に自動割り当てされます。`,
-    wontRun: '実行されません',
-    wontRunTip:
-      'Ready のカードはプロフィールが割り当てられて初めて実行されます。カードを開いて担当を設定するか、オーケストレーション設定でデフォルトの担当を設定してください。',
-    noHeartbeat: 'ハートビートなし',
-    expand: label => `${label} を展開`,
-    collapse: label => `${label} を折りたたむ`,
-    newTaskIn: label => `${label} に新しいタスク`,
-    empty: '空',
-    unassigned: '未割り当て',
-    filters: 'フィルタ',
-    allProfiles: 'すべてのプロフィール',
-    allTenants: 'すべてのテナント',
-    showArchived: 'アーカイブを表示',
-    groupRunning: '実行中をプロフィールでグループ化',
-    nSelected: n => `${n} 件選択中`,
-    moveToShort: '移動',
-    assign: '割り当て',
-    unassignAction: '割り当て解除',
-    archive: 'アーカイブ',
-    clearSelection: '選択をクリア（Esc）',
-    refused: '拒否されました',
-    bulkFailed: (failed, total, err) => `${total} 件中 ${failed} 件が失敗 — ${err}。失敗したカードは選択されたままです。`,
-    titlePlaceholderTriage: '大まかなアイデア — スペシファイアが具体化します',
-    titlePlaceholder: 'タイトル',
-    descPlaceholder: '説明（任意）',
-    priority: '優先度',
-    workspace: 'ワークスペース',
-    boardDefaultSuffix: '・ボード既定',
-    workspaceOverride: 'ワークスペースパス（任意の上書き）',
-    workspaceInherit: 'ボードのプロジェクトディレクトリを継承',
-    workspaceInheritDir: dir => `空欄にすると ${dir} を継承します`,
-    workspaceInheritGeneric: '空欄にするとボードのプロジェクトディレクトリを継承します。',
-    assignee: '担当',
-    defaultOption: name => `${name}（既定）`,
-    parkedOption: '未割り当て（保留 — 実行されません）',
-    skills: 'スキル（カンマ区切り）',
-    skillsPlaceholder: 'translation, github',
-    parent: '親（完了するまでブロック）',
-    noParent: '— 親なし —',
-    goalMode: 'ゴールモード（ジャッジが完了と認めるまでワーカーがループ）',
-    creating: '作成中…',
-    createTask: 'タスクを作成',
-    cancel: 'キャンセル',
-    save: '保存',
-    estimate: '見積もり',
-    estimateEffort: '工数を見積もり',
-    estimating: '見積もり中…',
-    reEstimate: '再見積もり',
-    makesModelCall: 'モデル呼び出しあり',
-    estimateTip: '補助モデルによるトークン数と複雑度の概算 — モデル呼び出しを行います。',
-    estimateTipLong: '補助モデルを呼び出してトークン数と複雑度を概算します。目安であり、請求ではありません。',
-    roughEstimate: '概算',
-    tokUnit: 'tok',
-    couldNotEstimate: '見積もりできませんでした',
-    complexity: { S: '小', M: '中', L: '大' },
-    introBody:
-      'カードはあなたではなくエージェントが実行します。担当を設定したカードを Ready に置くと、1分以内にエージェントが取得します。担当がなければ実行されません。トリアージ: エージェントがまずアイデアを適切なタスクに書き直します。Todo: 他のカード待ち。スケジュール: タイマー待ち。実行中とレビュー: エージェントのレーンなので手を出さないでください。ブロック: あなたの対応待ちです。結果はカードに戻ってきます。',
-    introGotIt: '了解',
-    evtCreated: (where, assignee) =>
-      `作成${where ? `（${where}）` : ''}${assignee ? `・${assignee} に割り当て` : ''}`,
-    evtMovedTo: col => `${col} へ移動`,
-    evtParentReopened: parent => `親 ${parent} が再オープン`,
-    evtAssignedTo: assignee => `${assignee} に割り当て`,
-    evtUnassigned: '割り当て解除',
-    evtCommentBy: author => `${author} のコメント`,
-    evtClaimedReview: 'レビューエージェントが取得',
-    evtClaimedWorker: 'ワーカーが取得',
-    evtWorkerStarted: 'ワーカー開始',
-    evtCompleted: '完了',
-    evtBlocked: 'ブロック — 人間の入力が必要',
-    evtUnblocked: col => `ブロック解除${col ? ` → ${col}` : ' → Ready'}`,
-    evtReclaimed: '再取得 — キューに戻しました',
-    evtSpecified: 'トリアージエージェントが仕様を作成',
-    evtPromoted: '依存関係が完了 — Ready に昇格',
-    evtScheduled: '後で実行するようスケジュール',
-    evtArchived: 'アーカイブ済み',
-    evtReprioritized: priority => `優先度を ${priority} に設定`,
-    someone: '誰か',
-    metaPriority: '優先度',
-    metaTenant: 'テナント',
-    metaCreatedBy: '作成者',
-    metaCreated: '作成',
-    metaWorkerPid: 'ワーカー PID',
-    readyUnassignedTitle: 'Ready ですが未割り当て — このカードは実行されません。',
-    readyUnassignedBody:
-      'ディスパッチャは担当のある Ready カードのみ取得します。上の担当フィールドでプロフィールを選ぶ（またはオーケストレーション設定でデフォルトの担当を設定する）と、1分以内に実行されます。',
-    diagnosticsN: n => `診断・${n}`,
-    commandCopied: 'コマンドをコピーしました',
-    description: '説明',
-    editDescription: '説明を編集',
-    cancelEdit: '編集をキャンセル',
-    noDescription: 'まだ説明はありません。',
-    result: '結果',
-    latestSummary: '最新のサマリー',
-    dependencies: '依存関係',
-    blockedBy: 'ブロック元',
-    blocks: 'ブロック先',
-    comments: n => `コメント・${n}`,
-    commentsHelpRunning:
-      'このタスクは実行中です。あなたのメモは数秒以内にワーカーの現在のターンに取り込まれます — ブロック/解除の操作は不要です。「メモを付けて再キュー」を選ぶと、メモを文脈に含めてタスクを最初からやり直します。',
-    commentsHelp: 'コメントはタスクのスレッドに追加されます。ワーカーがタスクを取得すると、文脈の一部として読み込みます。',
-    send: '送信',
-    comment: 'コメント',
-    messageWorker: '実行中のワーカーにメッセージ…',
-    addComment: 'コメントを追加…',
-    deliveredLive: '数秒以内に実行中のワーカーへ届きます。',
-    requeueWithNote: 'メモを付けて再キュー',
-    notePosted: 'メモを投稿しました — ワーカーを再キューしました',
-    activity: n => `アクティビティ・${n}`,
-    runs: n => `実行・${n}`,
-    workerLog: 'ワーカーログ',
-    workerLogTail: 'ワーカーログ・末尾',
-    attachments: n => `添付・${n}`,
-    noAttachments: 'まだ添付はありません。',
-    uploadAttachment: '添付をアップロード',
-    taskActions: 'タスクの操作',
-    copyTaskId: 'タスク ID をコピー',
-    copyTitle: 'タイトルをコピー',
-    copiedId: id => `${id} をコピーしました`,
-    copiedTitle: 'タイトルをコピーしました',
-    archiveTask: 'タスクをアーカイブ',
-    deleteTask: 'タスクを削除',
-    close: '閉じる',
-    working: '作業中',
-    board: 'ボード',
-    newBoard: '新しいボード',
-    newBoardDots: '新しいボード…',
-    boardSettings: 'ボード設定…',
-    boardSettingsFor: name => `ボード設定 — ${name}`,
-    name: '名前',
-    boardNamePlaceholder: 'ボード名',
-    slug: slug => `slug: ${slug}`,
-    project: 'プロジェクト',
-    noProject: 'プロジェクトなし（スクラッチのサンドボックス）',
-    projectHintPre:
-      '新しいタスクはプロジェクトのリポジトリで実行されます（タスクごとに worktree）。各タスクは作成時にワークスペースを上書きできます。プロジェクトの管理は ',
-    projectHintCmd: 'hermes project',
-    createBoard: 'ボードを作成',
-    orchestratorProfile: 'オーケストレータープロフィール',
-    defaultAssignee: 'デフォルトの担当',
-    defaultParen: '（既定）',
-    autoDecompose: 'トリアージタスクを自動分解',
-    profileDescriptions: 'プロフィールの説明',
-    profileDescriptionsHint:
-      '説明はデコンポーザーのルーティングを導きます。補助モデルで自動生成するか、自分で書いてください。',
-    profileGoodAt: 'このプロフィールの得意分野は？',
-    auto: '自動'
-  }
 })
