+++ apps/desktop/src/i18n/zh-hant.ts
@@ -2685,187 +2685,4 @@ export const zhHant = defineLocale({
       toggle: open => `${open ? '顯示' : '隱藏'}側邊欄`
     }
   },
-
-  kanban: {
-    nav: '看板',
-    openBoard: '看板：開啟面板',
-    countTip: (running, ready) => `看板 — 執行中 ${running}、就緒 ${ready}`,
-    col: {
-      triage: { label: '分類', help: '原始想法 — 由細化代理整理出規格。' },
-      todo: { label: '待辦', help: '等待相依項目，或未指派。' },
-      scheduled: { label: '已排程', help: '等待預定時間到來。' },
-      ready: { label: '就緒', help: '相依項目已滿足 — 指派一個設定檔，排程器便會執行它。' },
-      running: { label: '執行中', help: '已被工作單元領取 — 有代理在處理。由排程器設定。' },
-      blocked: { label: '受阻', help: '工作單元需要人工輸入。' },
-      review: { label: '審查', help: '審查代理正在檢查工作。由排程器設定。' },
-      done: { label: '完成', help: '已完成；相依它的子任務變為就緒。' },
-      archived: { label: '已封存', help: '從預設面板檢視中隱藏。' }
-    },
-    locked: {
-      review: '審查狀態由排程器在審查代理領取卡片時設定。',
-      running: '執行中由排程器在工作單元領取卡片時設定。',
-      scheduled: '排程需要喚醒時間 — 由代理設定；無法拖入。'
-    },
-    arcRunning: '有代理正在處理它。',
-    arcStale: '已領取，但超過 2 分鐘沒有工作單元心跳 — 排程器將重新領取。',
-    title: '看板',
-    orchestrationSettings: '編排設定',
-    newTask: '新增任務',
-    filterCards: '篩選卡片…',
-    noMatch: '沒有符合篩選條件的任務',
-    noTasks: '此面板尚無任務',
-    open: '開啟',
-    select: '選取（⌘點擊）',
-    deselect: '取消選取',
-    moveTo: label => `移至 ${label}`,
-    delete: '刪除',
-    reviewChecking: '審查代理正在檢查已完成的工作。',
-    attachedTip: name => `${name} 已接手 — 排程器將在下一個週期（≤1 分鐘）移交。`,
-    orchestratorTip: name => `${name}（編排者）將在下一個週期領取並撰寫規格。`,
-    autoAssignTip: name => `將在下一個排程週期自動指派給「${name}」（kanban.default_assignee）。`,
-    wontRun: '不會執行',
-    wontRunTip: '就緒卡片只有在指派了設定檔後才會執行。開啟卡片設定負責人，或在編排設定中設定預設負責人。',
-    noHeartbeat: '無心跳',
-    expand: label => `展開 ${label}`,
-    collapse: label => `摺疊 ${label}`,
-    newTaskIn: label => `在 ${label} 新增任務`,
-    empty: '空',
-    unassigned: '未指派',
-    filters: '篩選',
-    allProfiles: '所有設定檔',
-    allTenants: '所有租戶',
-    showArchived: '顯示已封存',
-    groupRunning: '依設定檔分組執行中',
-    nSelected: n => `已選取 ${n} 個`,
-    moveToShort: '移至',
-    assign: '指派',
-    unassignAction: '取消指派',
-    archive: '封存',
-    clearSelection: '清除選取（Esc）',
-    refused: '被拒絕',
-    bulkFailed: (failed, total, err) => `${total} 個中有 ${failed} 個失敗 — ${err}。失敗的卡片仍保持選取。`,
-    titlePlaceholderTriage: '大致想法 — 細化代理會補全',
-    titlePlaceholder: '標題',
-    descPlaceholder: '描述（選填）',
-    priority: '優先順序',
-    workspace: '工作區',
-    boardDefaultSuffix: '・面板預設',
-    workspaceOverride: '工作區路徑（選填覆寫）',
-    workspaceInherit: '繼承面板的專案目錄',
-    workspaceInheritDir: dir => `留空則繼承 ${dir}`,
-    workspaceInheritGeneric: '留空則繼承面板的專案目錄。',
-    assignee: '負責人',
-    defaultOption: name => `${name}（預設）`,
-    parkedOption: '未指派（擱置 — 不會執行）',
-    skills: '技能（以逗號分隔）',
-    skillsPlaceholder: 'translation, github',
-    parent: '父任務（完成前會阻擋）',
-    noParent: '— 無父任務 —',
-    goalMode: '目標模式（工作單元循環直到評判代理認可完成）',
-    creating: '建立中…',
-    createTask: '建立任務',
-    cancel: '取消',
-    save: '儲存',
-    estimate: '估算',
-    estimateEffort: '估算工作量',
-    estimating: '估算中…',
-    reEstimate: '重新估算',
-    makesModelCall: '會呼叫模型',
-    estimateTip: '由輔助模型對 token 數與複雜度的粗略估算 — 會呼叫模型。',
-    estimateTipLong: '快速呼叫輔助模型來估算 token 數與複雜度。僅供參考，並非帳單。',
-    roughEstimate: '粗略估算',
-    tokUnit: 'tok',
-    couldNotEstimate: '無法估算',
-    complexity: { S: '小', M: '中', L: '大' },
-    introBody:
-      '卡片不由你執行，而是由代理執行。把有負責人的卡片放入「就緒」，代理會在一分鐘內領取。沒有負責人就不會執行。分類：代理先把想法改寫成合適的任務。待辦：等待其他卡片。已排程：等待計時器。執行中與審查：這是代理的通道，請勿插手。受阻：正在等你。結果會回到卡片上。',
-    introGotIt: '知道了',
-    evtCreated: (where, assignee) =>
-      `已建立${where ? `（${where}）` : ''}${assignee ? `・指派給 ${assignee}` : ''}`,
-    evtMovedTo: col => `移至 ${col}`,
-    evtParentReopened: parent => `父任務 ${parent} 已重新開啟`,
-    evtAssignedTo: assignee => `指派給 ${assignee}`,
-    evtUnassigned: '取消指派',
-    evtCommentBy: author => `${author} 的留言`,
-    evtClaimedReview: '被審查代理領取',
-    evtClaimedWorker: '被工作單元領取',
-    evtWorkerStarted: '工作單元已啟動',
-    evtCompleted: '已完成',
-    evtBlocked: '受阻 — 需要人工輸入',
-    evtUnblocked: col => `已解除阻擋${col ? ` → ${col}` : ' → 就緒'}`,
-    evtReclaimed: '已重新領取 — 已放回佇列',
-    evtSpecified: '分類代理已撰寫規格',
-    evtPromoted: '相依項目已完成 — 提升為就緒',
-    evtScheduled: '已排程稍後執行',
-    evtArchived: '已封存',
-    evtReprioritized: priority => `優先順序設為 ${priority}`,
-    someone: '某人',
-    metaPriority: '優先順序',
-    metaTenant: '租戶',
-    metaCreatedBy: '建立者',
-    metaCreated: '建立於',
-    metaWorkerPid: '工作單元 PID',
-    readyUnassignedTitle: '就緒但未指派 — 這張卡片永遠不會執行。',
-    readyUnassignedBody:
-      '排程器只領取有負責人的就緒卡片。在上方的負責人欄位選擇一個設定檔（或在編排設定中設定預設負責人），它會在一分鐘內執行。',
-    diagnosticsN: n => `診斷・${n}`,
-    commandCopied: '指令已複製',
-    description: '描述',
-    editDescription: '編輯描述',
-    cancelEdit: '取消編輯',
-    noDescription: '尚無描述。',
-    result: '結果',
-    latestSummary: '最新摘要',
-    dependencies: '相依關係',
-    blockedBy: '受阻於',
-    blocks: '阻擋',
-    comments: n => `留言・${n}`,
-    commentsHelpRunning:
-      '此任務正在執行。你的備註會在幾秒內融入工作單元目前的回合 — 無需阻擋/解除操作。選擇「附上備註重新排入佇列」則會帶著你的備註從頭重跑任務。',
-    commentsHelp: '留言會加入任務討論串。工作單元領取任務時會將其作為脈絡的一部分讀取。',
-    send: '傳送',
-    comment: '留言',
-    messageWorker: '傳訊給執行中的工作單元…',
-    addComment: '新增留言…',
-    deliveredLive: '幾秒內送達執行中的工作單元。',
-    requeueWithNote: '附上備註重新排入佇列',
-    notePosted: '備註已發布 — 工作單元已重新排入佇列',
-    activity: n => `活動・${n}`,
-    runs: n => `執行・${n}`,
-    workerLog: '工作單元日誌',
-    workerLogTail: '工作單元日誌・末尾',
-    attachments: n => `附件・${n}`,
-    noAttachments: '尚無附件。',
-    uploadAttachment: '上傳附件',
-    taskActions: '任務操作',
-    copyTaskId: '複製任務 ID',
-    copyTitle: '複製標題',
-    copiedId: id => `已複製 ${id}`,
-    copiedTitle: '已複製標題',
-    archiveTask: '封存任務',
-    deleteTask: '刪除任務',
-    close: '關閉',
-    working: '進行中',
-    board: '面板',
-    newBoard: '新增面板',
-    newBoardDots: '新增面板…',
-    boardSettings: '面板設定…',
-    boardSettingsFor: name => `面板設定 — ${name}`,
-    name: '名稱',
-    boardNamePlaceholder: '面板名稱',
-    slug: slug => `slug: ${slug}`,
-    project: '專案',
-    noProject: '無專案（暫存沙箱）',
-    projectHintPre: '新任務將在專案的儲存庫中執行（每個任務一個 worktree）；每個任務在建立時仍可覆寫其工作區。管理專案請使用 ',
-    projectHintCmd: 'hermes project',
-    createBoard: '建立面板',
-    orchestratorProfile: '編排者設定檔',
-    defaultAssignee: '預設負責人',
-    defaultParen: '（預設）',
-    autoDecompose: '自動分解分類任務',
-    profileDescriptions: '設定檔說明',
-    profileDescriptionsHint: '說明用於引導分解器的路由。可用輔助模型自動產生，或自行填寫。',
-    profileGoodAt: '這個設定檔擅長什麼？',
-    auto: '自動'
-  }
 })
