+++ apps/desktop/src/i18n/zh.ts
@@ -3115,187 +3115,4 @@ export const zh: Translations = {
       toggle: open => `${open ? '显示' : '隐藏'}侧边栏`
     }
   },
-
-  kanban: {
-    nav: '看板',
-    openBoard: '看板：打开面板',
-    countTip: (running, ready) => `看板 — 运行中 ${running}、就绪 ${ready}`,
-    col: {
-      triage: { label: '分诊', help: '原始想法 — 由细化代理整理出规格。' },
-      todo: { label: '待办', help: '等待依赖，或未分配。' },
-      scheduled: { label: '已排期', help: '等待预定时间到来。' },
-      ready: { label: '就绪', help: '依赖已满足 — 分配一个配置档，调度器即会运行它。' },
-      running: { label: '运行中', help: '已被工作单元领取 — 有代理在处理。由调度器设置。' },
-      blocked: { label: '受阻', help: '工作单元需要人工输入。' },
-      review: { label: '审查', help: '审查代理正在检查工作。由调度器设置。' },
-      done: { label: '完成', help: '已完成；依赖它的子任务变为就绪。' },
-      archived: { label: '已归档', help: '从默认面板视图中隐藏。' }
-    },
-    locked: {
-      review: '审查状态由调度器在审查代理领取卡片时设置。',
-      running: '运行中由调度器在工作单元领取卡片时设置。',
-      scheduled: '排期需要唤醒时间 — 由代理设置；无法拖入。'
-    },
-    arcRunning: '有代理正在处理它。',
-    arcStale: '已领取，但超过 2 分钟没有工作单元心跳 — 调度器将重新领取。',
-    title: '看板',
-    orchestrationSettings: '编排设置',
-    newTask: '新建任务',
-    filterCards: '筛选卡片…',
-    noMatch: '没有符合筛选条件的任务',
-    noTasks: '此面板暂无任务',
-    open: '打开',
-    select: '选择（⌘点击）',
-    deselect: '取消选择',
-    moveTo: label => `移动到 ${label}`,
-    delete: '删除',
-    reviewChecking: '审查代理正在检查已完成的工作。',
-    attachedTip: name => `${name} 已接手 — 调度器将在下一个周期（≤1 分钟）移交。`,
-    orchestratorTip: name => `${name}（编排者）将在下一个周期领取并撰写规格。`,
-    autoAssignTip: name => `将在下一个调度周期自动分配给“${name}”（kanban.default_assignee）。`,
-    wontRun: '不会运行',
-    wontRunTip: '就绪卡片只有在分配了配置档后才会运行。打开卡片设置负责人，或在编排设置中配置默认负责人。',
-    noHeartbeat: '无心跳',
-    expand: label => `展开 ${label}`,
-    collapse: label => `折叠 ${label}`,
-    newTaskIn: label => `在 ${label} 新建任务`,
-    empty: '空',
-    unassigned: '未分配',
-    filters: '筛选',
-    allProfiles: '所有配置档',
-    allTenants: '所有租户',
-    showArchived: '显示已归档',
-    groupRunning: '按配置档分组运行中',
-    nSelected: n => `已选择 ${n} 个`,
-    moveToShort: '移动到',
-    assign: '分配',
-    unassignAction: '取消分配',
-    archive: '归档',
-    clearSelection: '清除选择（Esc）',
-    refused: '被拒绝',
-    bulkFailed: (failed, total, err) => `${total} 个中有 ${failed} 个失败 — ${err}。失败的卡片仍保持选中。`,
-    titlePlaceholderTriage: '大致想法 — 细化代理会补全',
-    titlePlaceholder: '标题',
-    descPlaceholder: '描述（可选）',
-    priority: '优先级',
-    workspace: '工作区',
-    boardDefaultSuffix: '・面板默认',
-    workspaceOverride: '工作区路径（可选覆盖）',
-    workspaceInherit: '继承面板的项目目录',
-    workspaceInheritDir: dir => `留空则继承 ${dir}`,
-    workspaceInheritGeneric: '留空则继承面板的项目目录。',
-    assignee: '负责人',
-    defaultOption: name => `${name}（默认）`,
-    parkedOption: '未分配（搁置 — 不会运行）',
-    skills: '技能（逗号分隔）',
-    skillsPlaceholder: 'translation, github',
-    parent: '父任务（完成前会阻塞）',
-    noParent: '— 无父任务 —',
-    goalMode: '目标模式（工作单元循环直到评判代理认可完成）',
-    creating: '创建中…',
-    createTask: '创建任务',
-    cancel: '取消',
-    save: '保存',
-    estimate: '估算',
-    estimateEffort: '估算工作量',
-    estimating: '估算中…',
-    reEstimate: '重新估算',
-    makesModelCall: '会调用模型',
-    estimateTip: '由辅助模型对令牌数和复杂度的粗略估算 — 会调用模型。',
-    estimateTipLong: '快速调用辅助模型来估算令牌数和复杂度。仅供参考，并非账单。',
-    roughEstimate: '粗略估算',
-    tokUnit: 'tok',
-    couldNotEstimate: '无法估算',
-    complexity: { S: '小', M: '中', L: '大' },
-    introBody:
-      '卡片不由你运行，而是由代理运行。把带有负责人的卡片放入“就绪”，代理会在一分钟内领取。没有负责人就不会运行。分诊：代理先把想法改写成合适的任务。待办：等待其他卡片。已排期：等待计时器。运行中与审查：这是代理的通道，请勿插手。受阻：正在等你。结果会回到卡片上。',
-    introGotIt: '知道了',
-    evtCreated: (where, assignee) =>
-      `已创建${where ? `（${where}）` : ''}${assignee ? `・分配给 ${assignee}` : ''}`,
-    evtMovedTo: col => `移动到 ${col}`,
-    evtParentReopened: parent => `父任务 ${parent} 已重新打开`,
-    evtAssignedTo: assignee => `分配给 ${assignee}`,
-    evtUnassigned: '取消分配',
-    evtCommentBy: author => `${author} 的评论`,
-    evtClaimedReview: '被审查代理领取',
-    evtClaimedWorker: '被工作单元领取',
-    evtWorkerStarted: '工作单元已启动',
-    evtCompleted: '已完成',
-    evtBlocked: '受阻 — 需要人工输入',
-    evtUnblocked: col => `已解除阻塞${col ? ` → ${col}` : ' → 就绪'}`,
-    evtReclaimed: '已重新领取 — 已放回队列',
-    evtSpecified: '分诊代理已撰写规格',
-    evtPromoted: '依赖已完成 — 提升为就绪',
-    evtScheduled: '已排期稍后运行',
-    evtArchived: '已归档',
-    evtReprioritized: priority => `优先级设为 ${priority}`,
-    someone: '某人',
-    metaPriority: '优先级',
-    metaTenant: '租户',
-    metaCreatedBy: '创建者',
-    metaCreated: '创建于',
-    metaWorkerPid: '工作单元 PID',
-    readyUnassignedTitle: '就绪但未分配 — 这张卡片永远不会运行。',
-    readyUnassignedBody:
-      '调度器只领取有负责人的就绪卡片。在上面的负责人字段选择一个配置档（或在编排设置中设置默认负责人），它会在一分钟内运行。',
-    diagnosticsN: n => `诊断・${n}`,
-    commandCopied: '命令已复制',
-    description: '描述',
-    editDescription: '编辑描述',
-    cancelEdit: '取消编辑',
-    noDescription: '暂无描述。',
-    result: '结果',
-    latestSummary: '最新摘要',
-    dependencies: '依赖关系',
-    blockedBy: '受阻于',
-    blocks: '阻塞',
-    comments: n => `评论・${n}`,
-    commentsHelpRunning:
-      '此任务正在运行。你的备注会在几秒内融入工作单元当前的回合 — 无需阻塞/解除操作。选择“附带备注重新入队”则会带着你的备注从头重跑任务。',
-    commentsHelp: '评论会添加到任务讨论串。工作单元领取任务时会将其作为上下文的一部分读取。',
-    send: '发送',
-    comment: '评论',
-    messageWorker: '给运行中的工作单元发消息…',
-    addComment: '添加评论…',
-    deliveredLive: '几秒内送达运行中的工作单元。',
-    requeueWithNote: '附带备注重新入队',
-    notePosted: '备注已发布 — 工作单元已重新入队',
-    activity: n => `活动・${n}`,
-    runs: n => `运行・${n}`,
-    workerLog: '工作单元日志',
-    workerLogTail: '工作单元日志・末尾',
-    attachments: n => `附件・${n}`,
-    noAttachments: '暂无附件。',
-    uploadAttachment: '上传附件',
-    taskActions: '任务操作',
-    copyTaskId: '复制任务 ID',
-    copyTitle: '复制标题',
-    copiedId: id => `已复制 ${id}`,
-    copiedTitle: '已复制标题',
-    archiveTask: '归档任务',
-    deleteTask: '删除任务',
-    close: '关闭',
-    working: '进行中',
-    board: '面板',
-    newBoard: '新建面板',
-    newBoardDots: '新建面板…',
-    boardSettings: '面板设置…',
-    boardSettingsFor: name => `面板设置 — ${name}`,
-    name: '名称',
-    boardNamePlaceholder: '面板名称',
-    slug: slug => `slug: ${slug}`,
-    project: '项目',
-    noProject: '无项目（临时沙箱）',
-    projectHintPre: '新任务将在项目的仓库中运行（每个任务一个 worktree）；每个任务在创建时仍可覆盖其工作区。管理项目请使用 ',
-    projectHintCmd: 'hermes project',
-    createBoard: '创建面板',
-    orchestratorProfile: '编排者配置档',
-    defaultAssignee: '默认负责人',
-    defaultParen: '（默认）',
-    autoDecompose: '自动分解分诊任务',
-    profileDescriptions: '配置档说明',
-    profileDescriptionsHint: '说明用于引导分解器的路由。可用辅助模型自动生成，或自行填写。',
-    profileGoodAt: '这个配置档擅长什么？',
-    auto: '自动'
-  }
 }
